backend/requirements.txt
flask==1.1.1
Pillow
flask-cors==3.0.8